Исправляет замечания наставника из прошлого ПР. Устанавливает redux и react-redux
