+++ package.json
@@ -34,6 +34,8 @@
     "prop-types": "15.7.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
-    "react-router-dom": "5.2.0"
+    "react-redux": "7.2.1",
+    "react-router-dom": "5.2.0",
+    "redux": "4.0.5"
   }
 }
