+++ src/components/tab-reviews/tab-reviews.test.js
@@ -1,51 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import TabReviews from "./tab-reviews.jsx";
-
-const reviews = [
-  {
-    id: 1,
-    author: `Kate Muir`,
-    date: `December 24, 2016`,
-    rating: `8.9`,
-    text: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years`,
-  },
-  {
-    id: 3,
-    author: `Bill Goodykoontz`,
-    date: `November 18, 2015`,
-    rating: `8.0`,
-    text: `Anderson's films are too precious for some, but for those of us willing to lose ourselves in them, they're a delight. "The Grand Budapest Hotel" is no different, except that he has added a hint of gravitas to the mix, improving the recipe.`,
-  },
-  {
-    id: 4,
-    author: `Amanda Greever`,
-    date: `November 18, 2015`,
-    rating: `8.0`,
-    text: `I didn't find it amusing, and while I can appreciate the creativity, it's an hour and 40 minutes I wish I could take back.`,
-  },
-  {
-    id: 4,
-    author: `Matthew Lickona`,
-    date: `December 20, 2016`,
-    rating: `7.2`,
-    text: `The mannered, madcap proceedings are often delightful, occasionally silly, and here and there, gruesome and/or heartbreaking.`,
-  },
-  {
-    id: 5,
-    author: `Paula Fleri-Soler`,
-    date: `December 20, 2016`,
-    rating: `7.6`,
-    text: `It is certainly a magical and childlike way of storytelling, even if the content is a little more adult`,
-  },
-  {
-    id: 6,
-    author: `Paula Fleri-Soler`,
-    date: `December 20, 2016`,
-    rating: `7.0`,
-    text: `It is certainly a magical and childlike way of storytelling, even if the content is a little more adult`,
-  },
-];
+import {reviews} from "../../mocks/testReviews.js";
 
 it(`Render TabReviews`, () => {
   const tree = renderer
