+++ src/components/tabs/tabs.test.js
@@ -1,9 +1,9 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import Tabs from "./tabs.jsx";
+import {Tab} from "../movie-card/movie-card.jsx";
 
-const tabs = [`Overview`, `Details`, `Reviews`];
-
+const tabs = Object.values(Tab);
 const activeTab = `Overview`;
 
 it(`Render Tabs`, () => {
