--- src/const.js
@@ -1,5 +1,0 @@
-export const Tab = {
-  OVERVIEW: `Overview`,
-  DETAILS: `Details`,
-  REVIEWS: `Reviews`,
-};
