+++ src/mocks/reviews.js
@@ -7,14 +7,14 @@ export default [
     text: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years`,
   },
   {
-    id: 3,
+    id: 2,
     author: `Bill Goodykoontz`,
     date: `November 18, 2015`,
     rating: `8.0`,
     text: `Anderson's films are too precious for some, but for those of us willing to lose ourselves in them, they're a delight. "The Grand Budapest Hotel" is no different, except that he has added a hint of gravitas to the mix, improving the recipe.`,
   },
   {
-    id: 4,
+    id: 3,
     author: `Amanda Greever`,
     date: `November 18, 2015`,
     rating: `8.0`,
