+++ src/mocks/testMoviesCards.js
@@ -1,0 +1,34 @@
+export const moviesCards = [
+  {
+    id: 1,
+    background: `img/bg-the-grand-budapest-hotel.jpg`,
+    imagePreview: `img/bohemian-rhapsody.jpg`,
+    poster: `https://placeimg.com/270/410/arch/grayscale`,
+    title: `Bohemian Rhapsody`,
+    genre: `Biography`,
+    date: `2018`,
+    description: `The story of the legendary British rock band Queen and lead singer Freddie Mercury`,
+    rating: `8.0`,
+    ratingCount: `305`,
+    director: `Bryan Singer`,
+    stars: [`Rami Malek`, `Lucy Boynton`, `Gwilym Lee`],
+    duration: `1h 39m`,
+    preview: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
+  },
+  {
+    id: 2,
+    background: `https://placeimg.com/1300/552/tech`,
+    imagePreview: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
+    poster: `https://placeimg.com/270/410/arch/grayscale`,
+    title: `Fantastic Beasts: The Crimes of Grindelwald`,
+    genre: `Drama`,
+    date: `2018`,
+    description: `The second installment of the "Fantastic Beasts" series featuring the adventures of Magizoologist Newt Scamander.`,
+    rating: `6.6`,
+    ratingCount: `423`,
+    director: `David Yates`,
+    stars: [`Eddie Redmayne`, `Katherine Waterston`, `Dan Fogler`],
+    duration: `1h 39m`,
+    preview: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
+  },
+];
